Apply firing policy on notification evaluation
Optimization would be to disable the trigger that have the
FIRE_ONCE_AFTER_N once it is fired.

We have to remove a bunch of const qualifier since the firing mechanism
requires that we do accounting inside the object. This accounting could
be done on the side. Storing it in the object is simply a convenience
for now.

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Change-Id: Id755b73c1f976a5a7d7a188656a1de21bd703143
Depends-on: lttng-ust: I5a800fc92e588c2a6a0e26282b0ad5f31c044479