trigger: introduce refcounting
A trigger object have a dual relationship with the action and condition
associated to it. On the API client side, the trigger does not own the
action and condition associated to him. In such case the internal flag
"owns_internal_objects" is set to false. This is the default.

The trigger take ownership of the action and condition only if it is
created from a buffer , `lttng_trigger_create_from_buffer`.

In such case, `owns_internal_objects` is set to true. The trigger object
is now responsible for the destruction of the action and condition
object.

Signed-off-by: Jonathan Rajotte <jonathan.rajotte-julien@efficios.com>
Change-Id: Ic9ee8753257fcbe15fc1321b18582628328c0154