common: move bytecode utilities from filter to its own file
We'll want to re-use the filter bytecode to implement the trigger event
rule condition field captures.  This is a preparatory patch that moves
some filter bytecode stuff in a location that is not filter-specific, so
it can be used for both filters and captures.

The content of common/filter/filter-bytecode.h is moved to
common/bytecode/bytecode.h.  Some declarations for the various bytecode
helpers are added to that file.  The implementation for these helpers is
moved from common/filter/filter-visitor-generate-bytecode.c to
common/bytecode/bytecode.c.

The content of src/common/bytecode is built as a library, so it can be
used by the filter-grammar-test program.

A following patch renames the content of bytecode/bytecode.h to remove
the "filter" part.

The rest of the changes is just to adapt the code to the changes
mentioned above.

Depends-on: lttng-ust: I5a800fc92e588c2a6a0e26282b0ad5f31c044479
Change-Id: Id602c9046bdc76791026c5b5a928387145d18e43
Signed-off-by: Simon Marchi <simon.marchi@efficios.com>
Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>