doc: rename "event rule" -> "recording event rule"
Starting from LTTng 2.13, a trigger can have an "event rule match"
condition. This condition contains an event rule, just like (logically)
what you create and enable with `lttng enable-event` does.

This patch updates the manual pages to disambiguate the "event rule"
term: lttng-add-trigger(1) uses "event rule" (the concept) while other
manual pages use "recording event rule", that is, an event rule of which
the matching action is to record the event to some channel.

As of LTTng 2.13, it is correct to say that a recording event rule
always belongs to a channel and that an "event rule match" condition
contains an event rule (of which the purpose is not to record to a
channel).

Signed-off-by: Philippe Proulx <eeppeliteloop@gmail.com>
Change-Id: I6ae3ad463b735594335e7511839bfe523a44f236